+++ karyon3-rxnetty/src/main/java/com/netflix/karyon/rxnetty/server/HttpRoutingRequestHandler.java
@@ -1,4 +1,4 @@
-package com.netflix.karyon.rxnetty;
+package com.netflix.karyon.rxnetty.server;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.handler.codec.http.HttpResponseStatus;
